moved patching to processing int an attempt to make reading and understanding both patching and preprocessing better.
